understand Request
This is a Node.js code that creates an HTTP server and listens on port 3000. When a client makes a request to the server, the request details ( request method, URL, and header information) are logged to the console, and the Node.js process is exited immediately by calling process.exit(). This means that the server will not actually send a response to the client, and the client will likely receive an error.