CODE
This is an example of a basic Node.js HTTP server that listens on port 3000 and handles incoming requests using the http.createServer() method.

When a request is received, the server checks the URL of the request. If the URL is "/", it responds with an HTML form that allows the user to enter a message and submit it to the server using a POST request. If the URL is not "/", it responds with a simple HTML page containing a heading.

The server also sets the Content-Type header of the response to text/html to indicate that the response body is HTML.

Note that the return res.end() statement is used to end the response and prevent any additional data from being sent to the client if the URL is "/". If the URL is not "/", the res.end() method is called at the end of the request handler to send the response to the client.