Add files via upload
This is a simple Express.js server written in JavaScript that creates two routes:

The first route, "/users", responds to HTTP GET requests by sending a list of users in an HTML unordered list.
The second route, "/", responds to HTTP GET requests by sending a simple HTML form to the client. This form allows the user to enter a message and submit it to the server using HTTP POST method.
Here's a brief overview of the code:

First, we import the "express" module and create an instance of the app.
Then we set the value of "PORT" to either the value of "process.env.PORT" or 5000 if "process.env.PORT" is not defined. This allows the server to listen on the correct port whether it is deployed on a hosting platform or running locally.
We define a middleware function for the "/users" route that sends an HTML response containing a list of users.
We define a middleware function for the "/" route that sends an HTML response containing a simple form for submitting messages.
Finally, we start the server listening on the specified port using the "listen" method of the app object and log a message to the console to indicate that the server has started.
Note: The server does not handle HTTP POST requests for the "/" route, so submitting the form will result in an error. To handle HTTP POST requests, we would need to define another middleware function for the "/" route that handles the form submission and sends a response.